# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FermeManager Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alimentation/repas:
    get:
      operationId: getRepas
      tags: [alimentation]
      summary: List meal records
      responses:
        "200":
          description: Meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repas"
    post:
      operationId: createRepas
      tags: [alimentation]
      summary: Record a meal distribution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepasBody"
      responses:
        "201":
          description: Meal recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repas"
  /alimentation/stocks:
    get:
      operationId: getStocks
      tags: [alimentation]
      summary: Get feed stock levels
      responses:
        "200":
          description: Stock levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stock"
    post:
      operationId: createStock
      tags: [alimentation]
      summary: Create a new feed stock entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockBody"
      responses:
        "201":
          description: Stock created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
  /alimentation/stocks/{id}:
    put:
      operationId: updateStock
      tags: [alimentation]
      summary: Update a stock entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockBody"
      responses:
        "200":
          description: Stock updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
    delete:
      operationId: deleteStock
      tags: [alimentation]
      summary: Delete a stock entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /alimentation/livraisons:
    get:
      operationId: getLivraisons
      tags: [alimentation]
      summary: List feed deliveries
      responses:
        "200":
          description: Deliveries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Livraison"
    post:
      operationId: createLivraison
      tags: [alimentation]
      summary: Record a feed delivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLivraisonBody"
      responses:
        "201":
          description: Delivery recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Livraison"
...
  /budget/categories:
    get:
      operationId: getBudgetCategories
      tags: [budget]
      summary: List budget categories
      responses:
        "200":
          description: Budget categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetCategorie"
    post:
      operationId: createBudgetCategorie
      tags: [budget]
      summary: Create a budget category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetCategorieBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategorie"
  /budget/depenses:
    get:
      operationId: getDepenses
      tags: [budget]
      summary: List all expenses
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Depense"
    post:
      operationId: createDepense
      tags: [budget]
      summary: Record an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Depense"
...
  /alimentation/repas/{id}:
    delete:
      operationId: deleteRepas
      tags: [alimentation]
      summary: Delete a meal record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /alimentation/livraisons/{id}:
    delete:
      operationId: deleteLivraison
      tags: [alimentation]
      summary: Delete a delivery record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /budget/categories/{id}:
    delete:
      operationId: deleteBudgetCategorie
      tags: [budget]
      summary: Delete a budget category and its expenses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /budget/depenses/{id}:
    delete:
      operationId: deleteDepense
      tags: [budget]
      summary: Delete an expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── UTILISATEURS (admin only) ───────────────────────────────────────────────
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Validate credentials and return the matching user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Utilisateur"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset a user's password by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Utilisateur"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Repas:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        heure:
          type: string
        batiment:
          type: string
        aliment:
          type: string
        quantiteDistribuee:
          type: number
        quantiteRefusee:
          type: number
        distribue_par:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, date, heure, batiment, aliment, quantiteDistribuee, quantiteRefusee, createdAt]
    CreateRepasBody:
      type: object
      properties:
        date:
          type: string
        heure:
          type: string
        batiment:
          type: string
        aliment:
          type: string
        quantiteDistribuee:
          type: number
        quantiteRefusee:
          type: number
        distribue_par:
          type: string
          nullable: true
      required: [date, heure, batiment, aliment, quantiteDistribuee, quantiteRefusee]
    Stock:
      type: object
      properties:
        id:
          type: integer
        aliment:
          type: string
        quantite:
          type: number
        capaciteMax:
          type: number
        updatedAt:
          type: string
      required: [id, aliment, quantite, capaciteMax, updatedAt]
    CreateStockBody:
      type: object
      properties:
        aliment:
          type: string
        quantite:
          type: number
        capaciteMax:
          type: number
      required: [aliment, quantite, capaciteMax]
    Livraison:
      type: object
      properties:
        id:
          type: integer
        fournisseur:
          type: string
        date:
          type: string
        aliment:
          type: string
        quantite:
          type: number
        prixTotal:
          type: number
          nullable: true
        qualite:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, fournisseur, date, aliment, quantite, createdAt]
    CreateLivraisonBody:
      type: object
      properties:
        fournisseur:
          type: string
        date:
          type: string
        aliment:
          type: string
        quantite:
          type: number
        prixTotal:
          type: number
          nullable: true
        qualite:
          type: string
          nullable: true
      required: [fournisseur, date, aliment, quantite]
...
    BudgetCategorie:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        budget:
          type: number
        depense:
          type: number
        couleur:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, nom, budget, depense, createdAt]
    CreateBudgetCategorieBody:
      type: object
      properties:
        nom:
          type: string
        budget:
          type: number
        depense:
          type: number
        couleur:
          type: string
          nullable: true
      required: [nom, budget, depense]
    Depense:
      type: object
      properties:
        id:
          type: integer
        categorieId:
          type: integer
        categorieNom:
          type: string
          nullable: true
        description:
          type: string
        montant:
          type: number
        date:
          type: string
        createdAt:
          type: string
      required: [id, categorieId, description, montant, date, createdAt]
    CreateDepenseBody:
      type: object
      properties:
        categorieId:
          type: integer
        description:
          type: string
        montant:
          type: number
        date:
          type: string
      required: [categorieId, description, montant, date]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Utilisateur:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        role:
          type: string
        modules:
          type: array
          items:
            type: string
        actif:
          type: boolean
        createdAt:
          type: string
      required: [id, username, nom, prenom, email, role, modules, actif, createdAt]
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    ResetPasswordBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]