# lib/api-zod/src/generated/types/utilisateur.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FermeManager Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface Utilisateur {
  id: number;
  username: string;
  nom: string;
  prenom: string;
  email: string;
  role: string;
  modules: string[];
  actif: boolean;
  createdAt: string;
}